Move all classes from enum to _core/enum
