+++ docs/api_reference/python-api-reference.md
@@ -33,7 +33,6 @@ are internal implementations.
 .. automodule:: power_grid_model.enum
    :undoc-members:
    :imported-members:
-   :exclude-members: IntEnum
    :show-inheritance:
 ```
